Database.HDBC.PostgreSQL.Pure

Description

This is a compatible interface with HDBC-postgresql's Database.HDBC.PostgreSQL except Config.

Prepared statements are closed when some requests come once Statements are GCed, because HDBC doesn't have "close" interface.

Connection

data Config

A configuration of a connection.

Default configuration is def, which is following.

>>> address def
AddressResolved 127.0.0.1:5432
>>> user def
"postgres"
>>> password def
""
>>> database def
""
>>> sendingBufferSize def
4096
>>> receptionBufferSize def
4096

encodeString def = \code -> case code of "UTF8" -> pure . fromString; _ -> const $ fail $ "unexpected character code: " <> show code
decodeString def = \code -> case code of "UTF8" -> pure . toString; _ -> const $ fail $ "unexpected character code: " <> show code

Constructors

Config

Fields

address :: Address

user :: String

password :: String

database :: String

sendingBufferSize :: Int

in byte

receptionBufferSize :: Int

in byte

encodeString :: ShortByteString -> StringEncoder

decodeString :: ShortByteString -> StringDecoder

data Connection

PostgreSQL connection.

data Address

IP address.

Constructors

AddressResolved SockAddr
Address which is DNS resolved.

AddressNotResolved HostName ServiceName
Address which is not DNS resolved.

withConnection :: Config -> (Connection -> IO a) -> IO a

Bracket function for a connection.

connect :: Config -> IO Connection

To connect to the server.

Transaction

begin :: Connection -> IO ()

To send BEGIN SQL statement.

Orphan instances

ToField Rational
Security risk of DoS attack.

You should convert Rational to Scientific with fromRationalRepetend in the user side. If the rational value is computed to repeating decimals like 1/3 = 0.3333., this consumes a lot of memories. This is provided because of the HDBC compatibility.

Instance details

Methods

toField :: MonadFail m => BackendParameters -> StringEncoder -> Maybe Oid -> FormatCode -> Rational -> m (Maybe ByteString)

ToField SqlValue

Instance details

Methods

toField :: MonadFail m => BackendParameters -> StringEncoder -> Maybe Oid -> FormatCode -> SqlValue -> m (Maybe ByteString)

FromField SqlValue

Instance details

Methods

fromField :: MonadFail m => StringDecoder -> ColumnInfo -> Maybe ByteString -> m SqlValue

Convertible Oid SqlTypeId

Instance details

Methods

safeConvert :: Oid -> ConvertResult SqlTypeId

ToField (TimeOfDay, TimeZone)

Instance details

Methods

toField :: MonadFail m => BackendParameters -> StringEncoder -> Maybe Oid -> FormatCode -> (TimeOfDay, TimeZone) -> m (Maybe ByteString)

FromField (TimeOfDay, TimeZone)

Instance details

Methods

fromField :: MonadFail m => StringDecoder -> ColumnInfo -> Maybe ByteString -> m (TimeOfDay, TimeZone)