Database.PostgreSQL.Pure

Description

This is a client library for PostgreSQL Database which has following features.

faster and less CPU load

especially on multi-core environments

pure Haskell implementations

no libpq dependency

easy to build even on Windows

implements extended query protocol

about extended query protocol, see https://www.postgresql.org/docs/current/protocol-flow.html#PROTOCOL-FLOW-EXT-QUERY

Typical Example

Prepare a following table.

CREATE TABLE person (
  id serial PRIMARY KEY,
  name varchar(255) NOT NULL
);
INSERT INTO person (name) VALUES ('Ada');

You can run like following to get the record whose ID is 1.

>>> :set -XOverloadedStrings
>>> :set -XFlexibleContexts
>>> :set -XDataKinds
>>> :set -XTypeFamilies
>>> :set -XTypeApplications
>>> 
>>> import Database.PostgreSQL.Pure
>>> import Data.Default.Class (def)
>>> import Data.Int (Int32)
>>> import Data.ByteString (ByteString)
>>> import Data.Tuple.Only (Only (Only))
>>> import Data.Tuple.Homotuple.Only ()
>>> import Data.Maybe (fromMaybe)
>>> import System.Environment (lookupEnv)
>>> 
>>> getEnvDef name value = fromMaybe value <$> lookupEnv name
>>> 
>>> host' <- getEnvDef "PURE_HOST" "127.0.0.1"
>>> port' <- getEnvDef "PURE_PORT" "5432"
>>> user' <- getEnvDef "PURE_USER" "postgres"
>>> password' <- getEnvDef "PURE_PASSWORD" ""
>>> database' <- getEnvDef "PURE_DATABASE" "postgres"
>>> 
>>> conn <- connect def { address = AddressNotResolved host' port', user = user', password = password', database = database' }
>>> preparedStatementProcedure = parse "" "SELECT id, name FROM person WHERE id = $1" Nothing
>>> portalProcedure <- bind @_ @2 @_ @_ "" BinaryFormat BinaryFormat (parameters conn) (const $ fail "") (Only (1 :: Int32)) preparedStatementProcedure
>>> executedProcedure = execute @_ @_ @(Int32, ByteString) 0 (const $ fail "") portalProcedure
>>> ((_, _, e, _), _) <- sync conn executedProcedure
>>> records e
[(1,"Ada")]

Connection

data Config

A configuration of a connection.

Default configuration is def, which is following.

>>> address def
AddressResolved 127.0.0.1:5432
>>> user def
"postgres"
>>> password def
""
>>> database def
""
>>> sendingBufferSize def
4096
>>> receptionBufferSize def
4096

Constructors

Config

Fields

address :: Address

Server address.

user :: String

User name.

password :: String

Password of user.

database :: String

Database name.

sendingBufferSize :: Int

The size of sending buffer in byte.

receptionBufferSize :: Int

The size of receiving buffer in byte.

data Connection

PostgreSQL connection.

$sel:pid:Connection :: Connection -> Pid

The process ID of the server.

$sel:parameters:Connection :: Connection -> BackendParameters

Set of server parameters.

$sel:config:Connection :: Connection -> Config

Configuration of this connection.

data Address

IP address.

Constructors

AddressResolved SockAddr
Address which is DNS resolved.

AddressNotResolved HostName ServiceName
Address which is not DNS resolved.

type BackendParameters = Map ShortByteString ShortByteString

Set of server parameters.

type Pid = Int32

Proccess ID

withConnection :: Config -> (Connection -> IO a) -> IO a

Bracket function for a connection.

connect :: Config -> IO Connection

To connect to the server.

disconnect :: Connection -> IO ()

To disconnect to the server.

Extended Query

parse

Arguments

:: forall plen rlen. (KnownNat plen, KnownNat rlen, IsHomotupleItem plen Oid, IsHomotupleItem rlen ColumnInfo, IsHomotupleItem rlen Oid, IsHomolisttuple rlen Oid, IsHomolisttuple plen Oid, IsHomolisttuple rlen ColumnInfo)

=> PreparedStatementName
A new name of prepared statement.

-> Query
SQL whose placeoholder style is dollar style.

-> Maybe (Homotuple plen Oid, Homotuple rlen Oid)
On Nothing an additional pair of a request and a resposne is necessary. If concrete OIDs are given, it will be pass over.

-> PreparedStatementProcedure plen rlen

To get the procedure to build the message of parsing SQL query and to parse its response.

bind

Arguments

:: forall rlen param m. (Bind ps, ToRecord param, KnownNat rlen, HasLength (Homotuple rlen ColumnInfo), MonadFail m)

=> PortalName
A new name of portal.

-> FormatCode
Binary format or text format for the parameter.

-> FormatCode
Binary format or text format for the results.

-> BackendParameters
The set of the server parameters.

-> StringEncoder
How to encode strings.

-> param
Parameter for this query.

-> ps (Length param) rlen
Prepared statement.

-> m (PortalProcedure (Length param) rlen)

To get the procedure to build the message of binding the parameter and to parse its response.

execute

Arguments

:: forall plen result. (Execute p, FromRecord result, IsHomotupleItem (Length result) ColumnInfo, IsHomolisttuple (Length result) ColumnInfo)

=> Word
How many records to get. "0" means unlimited.

-> StringDecoder
How to decode strings.

-> p plen (Length result)
Portal.

-> ExecutedProcedure plen (Length result) result

To get the procedure to build the message of execution and to parse its response.

flush :: Message m => Connection -> m -> IO (MessageResult m)

To build and send the given message and a “Flush” message and to receive and parse those responses.

sync :: Message m => Connection -> m -> IO (MessageResult m, TransactionState)

To build and send the given message and a “Sync” message and to receive and parse those responses.

close :: Close p => p -> CloseProcedure

To build and send the “Close” message and to receive and parse its response.

data PreparedStatement (parameterLength :: Nat) (resultLength :: Nat)

This represents a prepared statement which is already processed by a server.

parameterLength is the number of columns of the parameter and resultLength is the number of columns of the results. This is the same with PreparedStatementProcedure, Portal, PortalProcedure, Executed and ExecutedProcedure.

data PreparedStatementProcedure (parameterLength :: Nat) (resultLength :: Nat)

This represents a prepared statement which is not yet processed by a server.

newtype PreparedStatementName

Name of a prepared statement.

Constructors

PreparedStatementName ByteString

data Portal (parameterLength :: Nat) (resultLength :: Nat)

This represents a portal which is already processed by a server.

data PortalProcedure (parameterLength :: Nat) (resultLength :: Nat)

This represents a portal which is not yet processed by a server.

newtype PortalName

Name of a portal.

Constructors

PortalName ByteString

data Executed (parameterLength :: Nat) (resultLength :: Nat) r

This represents a result of a Execute message which is already processed by a server.

data ExecutedProcedure (parameterLength :: Nat) (resultLength :: Nat) r

This represents a result of a Execute message which is not yet processed by a server.

data ExecuteResult

Result of a “Execute” message.

Constructors

ExecuteComplete CommandTag
All records gotten.

ExecuteEmptyQuery
No records.

ExecuteSuspended
Records are left yet.

data CloseProcedure

This represents a result of a “Close” message which is not yet processed by a server.

data CommandTag

Command tag, which means which SQL command is completed.

Constructors

InsertTag Oid Int

DeleteTag Int

UpdateTag Int

SelectTag Int

MoveTag Int

FetchTag Int

CopyTag Int

CreateTableTag

DropTableTag

BeginTag

CommitTag

RollbackTag

SetTag

newtype Query

SQL query.

This fromString counts only ASCII, becouse it is the same with ByteString.

Constructors

Query ByteString

data FormatCode

Format code of parameters of results.

Constructors

TextFormat

BinaryFormat

data ColumnInfo

Metadata of a column.

class Message m

This means that r is a objective of flush and sync.

type family MessageResult m :: Type

To convert a type which means that it is not processed by the server to a respective type which means that it is processed by the server.

class Bind ps

This means that ps is a objective of bind.

Minimal complete definition

bind

class Execute p

This means that p is a objective of execute.

Minimal complete definition

execute

class Close p

This means that p is a objective of close.

Minimal complete definition

close

type StringEncoder = String -> Either String ByteString

Encoder of strings which may fail.

type StringDecoder = ByteString -> Either String String

Decoder of strings which may fail.

class HasName r

This means that r has a name accesser.

type family Name r :: Type

Type of name of r.

class HasParameterOids r a

This means that r has a parameterOids accesser.

name :: HasName r => r -> Name r

To get a name of r.

parameterOids :: HasParameterOids r a => r -> a

To get OIDs of a parameter.

resultInfos :: (IsHomolisttuple m ColumnInfo, IsHomotupleItem m ColumnInfo) => PreparedStatement n m -> Homotuple m ColumnInfo

To get a list of column infos of the result record.

result :: Executed n m r -> ExecuteResult

To get the result of Executed.

records :: Executed n m r -> [r]

To get the records of Executed.

Transaction

begin :: ExecutedProcedure 0 0 ()

To send BEGIN SQL statement.

commit :: ExecutedProcedure 0 0 ()

To send COMMIT SQL statement.

rollback :: ExecutedProcedure 0 0 ()

To send ROLLBACK SQL statement.

data TransactionState

Transaction state of a server.

Constructors

Idle
Not in a transaction block.

Block
In a transaction block.

Failed
Transaction failed.

Record

class FromField a where

This means that a field can be decoded as a.

Methods

fromField :: MonadFail m => StringDecoder -> ColumnInfo -> Maybe ByteString -> m a

Decoder of a field.

class FromRecord a where

This means that a record can be parsed as a.

Minimal complete definition

Nothing

Methods

fromRecord :: StringDecoder -> [ColumnInfo] -> Parser a

Decoder of a record.

default fromRecord :: (Generic a, GFromRecord (Rep a)) => StringDecoder -> [ColumnInfo] -> Parser a

class ToField a where

This means that a can be encoded to a field.

Methods

toField :: MonadFail m => BackendParameters -> StringEncoder -> Maybe Oid -> FormatCode -> a -> m (Maybe ByteString)

Encoder of a field.

class ToRecord a where

This means that a can be encoded to a record.

Minimal complete definition

Nothing

Methods

toRecord :: MonadFail m => BackendParameters -> StringEncoder -> Maybe [Oid] -> [FormatCode] -> a -> m [Maybe ByteString]

Encoder of a field.

default toRecord :: (MonadFail m, Generic a, GToRecord (Rep a)) => BackendParameters -> StringEncoder -> Maybe [Oid] -> [FormatCode] -> a -> m [Maybe ByteString]

data Raw where

Data without encoding nor decoding of a field.

Bundled Patterns

pattern Value :: ByteString -> Raw
Not NULL.

pattern Null :: Raw
NULL.

type family Length a :: Nat

The number of columns.

Exception

data Exception

Root exception.

Exception
  ├ ErrorResponse
  └ ResponseParsingFailed

Constructors

forall e.Exception e => Exception e

data ErrorResponse

This means that the server responds an error.

Constructors

ErrorResponse

Fields

severity :: ByteString

code :: ByteString

message :: ByteString

transactionState :: Maybe TransactionState

newtype ResponseParsingFailed

This means that the server responds an unknown message.

Constructors

ResponseParsingFailed

Fields

causedBy :: String

OID

data Oid

OID.

Constant values are listed in Database.PostgreSQL.Pure.Oid.